+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]## [0.5.4.2b]
+### Added
+- MapManager.get_tooltip_table to get a tooltip formatted as table
 
 ## [0.5.4.1] - 2022-01-24
 ### Changed